requirements.txt
Flask==1.1.2
Flask-Env==2.0.0
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.11.0
marshmallow-sqlalchemy==0.22.3
FLask-Script==2.0.6

sentry_sdk[flask]==0.14.3

# test
pytest==5.4.1
pytest-flask-sqlalchemy==1.0.2
freezegun==0.3.15
